Add rewrite rule for date_part('year', X) -> year(X) and add more date part tests

+++ src/function/scalar/date/year.cpp
@@ -15,18 +15,18 @@ struct YearOperator {
 	}
 };
 
-template <> int32_t YearOperator::Operation(date_t input) {
+template <> int64_t YearOperator::Operation(date_t input) {
 	return Date::ExtractYear(input);
 }
 
-template <> int32_t YearOperator::Operation(timestamp_t input) {
+template <> int64_t YearOperator::Operation(timestamp_t input) {
 	return Date::ExtractYear(Timestamp::GetDate(input));
 }
 
 void YearFun::RegisterFunction(BuiltinFunctions &set) {
 	ScalarFunctionSet year("year");
-	year.AddFunction(ScalarFunction({SQLType::DATE}, SQLType::INTEGER, ScalarFunction::UnaryFunction<date_t, int32_t, YearOperator>));
-	year.AddFunction(ScalarFunction({SQLType::TIMESTAMP}, SQLType::INTEGER, ScalarFunction::UnaryFunction<timestamp_t, int32_t, YearOperator>));
+	year.AddFunction(ScalarFunction({SQLType::DATE}, SQLType::BIGINT, ScalarFunction::UnaryFunction<date_t, int64_t, YearOperator>));
+	year.AddFunction(ScalarFunction({SQLType::TIMESTAMP}, SQLType::BIGINT, ScalarFunction::UnaryFunction<timestamp_t, int64_t, YearOperator>));
 	set.AddFunction(year);
 }
